+++ src/resources/js/Pages/Home.vue
@@ -1,0 +1,18 @@
+<script setup>
+import { Head, Link } from '@inertiajs/vue3';
+
+defineProps({
+
+});
+</script>
+
+<template>
+    <Head title="Social Media Website" />
+
+    Test
+
+</template>
+
+<style>
+
+</style>